+++ source/extensions/omni.isaac.sensor/docs/CHANGELOG.md
@@ -40,6 +40,7 @@
 ## [9.6.0] - 2023-09-25
 ### Added
 - depthRange output to IsaacReadRTXLidarData
+- numBeams output to IsaacReadRTXLidarData
 
 ## [9.5.1] - 2023-09-20
 ### Fixed